+++ src/utils.ts
@@ -1,3 +1,5 @@
+import * as actionsCore from '@actions/core'
+
 export function maxBy<T>(array: T[], selector: (element: T) => number): T {
   if (array.length === 0) {
     throw new Error('Array empty')
@@ -13,4 +15,18 @@ export function maxBy<T>(array: T[], selector: (element: T) => number): T {
     }
   }
   return maxElement
+}
+
+export function parseSuccessConclusions(successConclusions: string, core: typeof actionsCore): string[] | undefined {
+  const regex =
+    /^(success|failure|neutral|cancelled|skipped|timed_out|action_required)(\|(success|failure|neutral|cancelled|skipped|timed_out|action_required))*$/
+  if (!regex.test(successConclusions)) {
+    core.setFailed(
+      "Invalid 'successConclusions'. It must be a pipe-separated non-empty subset of the options 'success|failure|neutral|cancelled|skipped|timed_out|action_required'"
+    )
+    return undefined
+  }
+  const result = successConclusions.split('|')
+  core.debug(`successConclusions.split('|'): ${JSON.stringify(result)}`)
+  return result
 }
